providers: serverscom: implement conversion rules
This adds converion rules for Servers.com hosting CSV reports.
Known issues:
1. Tags are to be implemented in converter
2. CommitmentDiscountCategory doesn't pass validation agains spec 1.0
   because of a bug in validator (asserts to be not null always when the
   field is nullable in certain case when CommitmentDiscountId is null)
3. ResourceType dimension from spec 1.0 is not yet supported by converter

This provided provides csv in a bad format that requires a lot of
converions hardly possible by means of focus-converter so its configs
come with a handy prepare_data.py that takes provider csv and prepares
it for conversion in a simpler way it could.

Signed-off-by: Stanislav Hanzhin-Tsvetkov <s.hanzhin@gaijin.team>